+++ src/main/java/com/nearby/shops/repositories/AbonnementRepository.java
@@ -3,6 +3,6 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Abonnement;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AbonnementRepository extends CrudRepository<Abonnement,Integer> {
+public interface AbonnementRepository extends CrudRepository<Abonnement,Long> {
 
 }
