+++ src/main/java/com/nearby/shops/repositories/FactureRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Facture;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FactureRepository extends CrudRepository<Facture,Integer> {
+public interface FactureRepository extends CrudRepository<Facture,Long> {
 }
