+++ src/main/java/com/nearby/shops/repositories/LigneCommandeRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.LigneCommande;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LigneCommandeRepository extends CrudRepository<LigneCommande,Integer> {
+public interface LigneCommandeRepository extends CrudRepository<LigneCommande,Long> {
 }
