+++ src/main/java/com/nearby/shops/repositories/LocalisationRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Localisation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LocalisationRepository extends CrudRepository<Localisation,Integer> {
+public interface LocalisationRepository extends CrudRepository<Localisation,Long> {
 }
